SAT backends build system refactoring (WIP...)
