+++ tests/CMakeLists.txt
@@ -39,7 +39,8 @@ set(
 )
 
 add_executable(unit_tests ${UNIT_TESTS})
-target_link_libraries(unit_tests PRIVATE black Catch2::Catch2)
+target_link_library_with_global_ctors(unit_tests PRIVATE black)
+target_link_libraries(unit_tests PRIVATE Catch2::Catch2)
 target_enable_warnings(unit_tests)
 target_code_coverage(unit_tests)
 add_sanitizers(unit_tests)
